Changed install script.
Change the installation process to copy each file separately instead of
usint tar. For each file we determine if the file has changed and opnly
copy if the file has changed.

Currently we copy entire subdirectories without processing. The next
step is to do the file check for each contained file.